requirements.txt
# Add packages you need here, one package per line
numpy
biopython